backend/requirements.txt
appdirs==1.4.3
astroid==2.4.2
boto3==1.14.33
botocore==1.17.33
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
docutils==0.15.2
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pylint==2.5.3
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.14.0
pytoml==0.1.21
redis==3.5.3
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
rasa==1.10.3
atlassian-python-api==1.16.0
